added comments for input data

+++ examples/xgboost/methods/guest/src/main.rs
@@ -24,16 +24,20 @@ pub fn main() {
     // In order for the trained model to make a prediction using the input data, the input data must be formatted as a forust_ml::Matrix type.
     // Note:  We cannot send the data from the host to the guest as a Matrix struct because the forust-ml crate does not enable serialization 
     // for Matrix objects.
+
     // To construct a Matrix object, we need three inputs:
-    // The first input is the raw input data that we read in from the host.  
+    // The first input is the raw input data.  
     // The second value is the number of instances, which in this case is one. 
     // The third value is the number of variables per instance, which in our case is two (block number and number of transactions in a block).
     // Note: You will need to adjust the second and third inputs depending on how many instances you pass the model and how many variables 
     // comprise each instance.
+    //*************************************************//
+    // MODIFY THE MATRIX TO MATCH YOUR INPUT DATA HERE //
+    //*************************************************//
     let matrix = Matrix::new(&input, 1, 2);
 
     // We call the predict() function on the trained XGBoost model to perform inference
-    let answer: f32 = xgboost_model.predict(&matrix, true)[0] as f32; 
+    let answer: f64 = xgboost_model.predict(&matrix, true)[0]; 
 
     // write public output to the journal
     env::commit(&answer);
